# swagger.yaml
openapi: 3.0.0
info:
  title: Tuk-Tuk Tracker API
  version: 1.0.0
  description: |
    ## Real-Time Three-Wheeler (Tuk-Tuk) Tracking & Movement Logging System
    Developed for **Sri Lanka Police** to support operational visibility and investigations.

  contact:
    name: Sri Lanka Police IT Division
  license:
    name: MIT
servers:
  - url: /api/v1
    description: API Version 1

#security scheme
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Enter your JWT token obtained from POST /auth/login.
        Format: `Bearer <token>`

#these are reusable schemas
  schemas:
...
    LocationPing:
      type: object
      properties:
        _id:
          type: string
          example: 64a1f2b3c4d5e6f7a8b9c0d6
        vehicle:
          type: string
          description: MongoDB ObjectId of the vehicle
          example: 64a1f2b3c4d5e6f7a8b9c0d5
        latitude:
          type: number
          format: float
          example: 6.9271
        longitude:
          type: number
          format: float
          example: 79.8612
        speed:
          type: number
          example: 42
        heading:
          type: number
          description: Direction in degrees 0-360
          example: 180
        accuracy:
          type: number
          description: GPS accuracy in metres
          example: 5.2
        timestamp:
          type: string
          format: date-time
          example: "2025-04-24T10:23:00.000Z"
        isAnomaly:
          type: boolean
          description: True when this ping was flagged as suspicious by the pattern engine
          example: false
        anomaly:
          type: string
          nullable: true
          description: The specific anomaly pattern detected
          enum:
            - NIGHT_MOVEMENT
            - SPEEDING
            - BOUNDARY_CROSS
            - STATIONARY
            - ERRATIC
          example: null
...
    SuspiciousVehicle:
      type: object
      description: |
        A vehicle that has triggered one or more anomaly patterns.
        Returned by the GET /alerts/suspicious endpoint.
        Sorted by severityScore descending — most suspicious first.
      properties:
        vehicleId:
          type: string
          description: MongoDB ObjectId of the vehicle
          example: 64a1f2b3c4d5e6f7a8b9c0d5
        registrationNumber:
          type: string
          description: Human-readable vehicle registration
          example: REG101
        vehicleStatus:
          type: string
          description: Current operational status of the vehicle
          enum:
            - active
            - inactive
            - flagged
          example: active
        vehicleColor:
          type: string
          example: Yellow
        assignedDistrict:
          type: string
          description: Name of the district this vehicle is registered to
          example: Colombo
        latestAnomalyType:
          type: string
          description: The most recently detected anomaly type for this vehicle
          enum:
            - NIGHT_MOVEMENT
            - SPEEDING
            - BOUNDARY_CROSS
            - STATIONARY
            - ERRATIC
          example: SPEEDING
        anomalyTypes:
          type: array
          description: All unique anomaly types this vehicle has triggered
          items:
            type: string
          example:
            - SPEEDING
            - NIGHT_MOVEMENT
        totalAnomalyCount:
          type: integer
          description: Total number of anomalous pings recorded for this vehicle
          example: 24
        severityScore:
          type: integer
          description: |
            Number of distinct anomaly types triggered.
            Higher score = more suspicious = appears higher in results.
            Score of 1 = one pattern type, Score of 3 = three different patterns.
          example: 2
        lastSeenAt:
          type: string
          format: date-time
          description: Timestamp of the most recent anomalous ping
          example: "2025-04-24T10:23:00.000Z"
        lastKnownLocation:
          type: object
          description: GPS coordinates at the time of the most recent anomaly
          properties:
            latitude:
              type: number
              example: 6.9271
            longitude:
              type: number
              example: 79.8612
        lastSpeed:
          type: number
          description: Speed recorded at the most recent anomalous ping
          example: 95
    AlertSummary:
      type: object
      description: Count of vehicles that triggered each anomaly type
      properties:
        NIGHT_MOVEMENT:
          type: integer
          description: Vehicles detected operating between midnight and 5am
          example: 8
        SPEEDING:
          type: integer
          description: Vehicles detected exceeding 80 km/h
          example: 8
        BOUNDARY_CROSS:
          type: integer
          description: Vehicles detected outside their assigned district
          example: 8
        STATIONARY:
          type: integer
          description: Vehicles detected stopped in one location too long
          example: 8
        ERRATIC:
          type: integer
          description: Vehicles detected with irregular zigzag movement
          example: 8
    AlertsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        totalSuspiciousVehicles:
          type: integer
          description: Total number of vehicles with at least one anomaly
          example: 40
        anomalySummary:
          $ref: '#/components/schemas/AlertSummary'
        filters:
          type: object
          description: The filters that were applied to this response
          properties:
            type:
              type: string
              description: Anomaly type filter applied (or "all" if none)
              example: all
            since:
              type: string
              description: Time filter applied (or "all time" if none)
              example: all time
            scope:
              type: string
              description: The role of the user who made the request
              example: hq_admin
        data:
          type: array
          description: Suspicious vehicles sorted by severity score descending
          items:
            $ref: '#/components/schemas/SuspiciousVehicle'

    #these are the common responses
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Description of what went wrong
...
    PaginatedLocations:
      type: object
      properties:
        success:
          type: boolean
          example: true
        results:
          type: integer
          example: 5
        data:
          type: array
          items:
            $ref: '#/components/schemas/LocationPing'

  #these are reusable params
  parameters:
    pageParam:
      in: query
      name: page
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number for pagination
      example: 1
    limitParam:
      in: query
      name: limit
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
      description: Number of results per page
      example: 10
...
  responses:
    Unauthorized:
      description: No token provided or token is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Access denied. Please log in to continue.
    Forbidden:
      description: Token valid but role does not have permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Your role 'station_officer' does not have permission to perform this action.
...
security:
  - bearerAuth: []

#tags for each section
...
paths:
...
  /locations/latest:
    get:
      summary: Get latest location of all vehicles
      description: |
        Returns the most recent GPS ping for every vehicle.

        **Allowed roles:** `hq_admin`, `provincial_officer`, `station_officer`
      operationId: getAllLatestLocations
      tags:
        - Tracking
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: Latest ping per vehicle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedLocations'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /locations/district/{districtName}:
    get:
      summary: Get latest location of ALL vehicles in a district
      description: |
        Returns the most recent GPS ping for every vehicle in the district.
        **Allowed roles:** `hq_admin`, `provincial_officer`, `station_officer`
      operationId: getDistrictLatest
      tags:
        - Tracking
      parameters:
        - in: path
          name: districtName
          required: true
          schema:
            type: string
          example: Colombo
      responses:
        '200':
          description: Latest ping per vehicle in district
          content:
            application/json:
              example:
                success: true
                count: 45
                data:
                  - vehicle: 64a1f2b3c4d5e6f7a8b9c0d5
                    latitude: 6.9271
                    longitude: 79.8612
                    speed: 42
                    timestamp: "2025-04-24T10:23:00.000Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /locations/{regNo}/latest:
    get:
      summary: Live view — latest location of a specific vehicle
      description: |
        Returns the single most recent GPS ping.
        Data is never more than 10 seconds old while simulate.js is running.
      operationId: getLatestLocation
      tags:
        - Tracking
      parameters:
        - in: path
          name: regNo
          required: true
          schema:
            type: string
          example: REG101
      responses:
        '200':
          description: Most recent location ping
          content:
            application/json:
              example:
                success: true
                data:
                  vehicle:
                    _id: 64a1f2b3c4d5e6f7a8b9c0d5
                    registrationNumber: REG101
                  latitude: 6.9271
                  longitude: 79.8612
                  speed: 42
                  timestamp: "2025-04-24T10:23:00.000Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No location data found for this vehicle
  /locations/{regNo}/history:
    get:
      summary: Movement log — GPS history within a time window
      description: |
        Returns all pings between optional `from` and `to` timestamps.
        Sorted newest first.
      operationId: getHistory
      tags:
        - Tracking
      parameters:
        - in: path
          name: regNo
          required: true
          schema:
            type: string
          example: REG101
        - in: query
          name: from
          schema:
            type: string
            format: date-time
          example: "2025-04-18T00:00:00Z"
        - in: query
          name: to
          schema:
            type: string
            format: date-time
          example: "2025-04-24T23:59:59Z"
      responses:
        '200':
          description: Array of location pings in time window
          content:
            application/json:
              example:
                success: true
                results: 42
                data:
                  - latitude: 6.9271
                    longitude: 79.8612
                    speed: 42
                    timestamp: "2025-04-24T10:23:00.000Z"
        '401':
          $ref: '#/components/responses/Unauthorized'

  #suspicious tuk tuk detection
  /alerts/suspicious:
    get:
      summary: Get all suspicious vehicles with anomaly patterns
      description: |
        Returns vehicles that have triggered anomaly flags from the
        pattern detection engine.
      operationId: getSuspiciousVehicles
      tags:
        - Alerts
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum:
              - NIGHT_MOVEMENT
              - SPEEDING
              - BOUNDARY_CROSS
              - STATIONARY
              - ERRATIC
          description: Filter by specific anomaly type
          example: SPEEDING
        - in: query
          name: since
          schema:
            type: string
            format: date-time
          description: |
            Only return anomalies detected after this timestamp.
          example: "2025-04-20T00:00:00Z"
      responses:
        '200':
          description: Suspicious vehicles grouped and sorted by severity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertsResponse'
              examples:
                all_suspicious:
                  summary: All suspicious vehicles (no filter)
                  value:
                    success: true
                    totalSuspiciousVehicles: 40
                    anomalySummary:
                      NIGHT_MOVEMENT: 8
                      SPEEDING: 8
                      BOUNDARY_CROSS: 8
                      STATIONARY: 8
                      ERRATIC: 8
                    filters:
                      type: all
                      since: all time
                      scope: hq_admin
                    data:
                      - vehicleId: 64a1f2b3c4d5e6f7a8b9c0d5
                        registrationNumber: REG101
                        vehicleStatus: active
                        vehicleColor: Yellow
                        assignedDistrict: Colombo
                        latestAnomalyType: SPEEDING
                        anomalyTypes:
                          - SPEEDING
                          - NIGHT_MOVEMENT
                        totalAnomalyCount: 24
                        severityScore: 2
                        lastSeenAt: "2025-04-24T10:23:00.000Z"
                        lastKnownLocation:
                          latitude: 6.9271
                          longitude: 79.8612
                        lastSpeed: 95
                      - vehicleId: 64a1f2b3c4d5e6f7a8b9c0d6
                        registrationNumber: REG116
                        vehicleStatus: active
                        vehicleColor: Blue
                        assignedDistrict: Kandy
                        latestAnomalyType: BOUNDARY_CROSS
                        anomalyTypes:
                          - BOUNDARY_CROSS
                        totalAnomalyCount: 11
                        severityScore: 1
                        lastSeenAt: "2025-04-24T09:15:00.000Z"
                        lastKnownLocation:
                          latitude: 7.8500
                          longitude: 81.1200
                        lastSpeed: 32
                speeding_only:
                  summary: Only speeding vehicles
                  value:
                    success: true
                    totalSuspiciousVehicles: 8
                    anomalySummary:
                      NIGHT_MOVEMENT: 0
                      SPEEDING: 8
                      BOUNDARY_CROSS: 0
                      STATIONARY: 0
                      ERRATIC: 0
                    filters:
                      type: SPEEDING
                      since: all time
                      scope: hq_admin
                    data:
                      - registrationNumber: REG102
                        latestAnomalyType: SPEEDING
                        anomalyTypes:
                          - SPEEDING
                        totalAnomalyCount: 18
                        severityScore: 1
                        lastSpeed: 112
        '400':
          description: Invalid anomaly type provided in ?type parameter
          content:
            application/json:
              example:
                success: false
                message: "Invalid anomaly type. Valid types: NIGHT_MOVEMENT, SPEEDING, BOUNDARY_CROSS, STATIONARY, ERRATIC"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /alerts/suspicious/district/{districtName}:
    get:
      summary: Get all suspicious vehicles in a district
      description: |
        Returns vehicles that have triggered anomaly flags from the
        pattern detection engine.
      operationId: getSuspiciousVehiclesByDistrict
      tags:
        - Alerts
      parameters:
        - in: path
          name: districtName
          required: true
          schema:
            type: string
          example: Colombo
      responses:
        '200':
          description: Suspicious vehicles in district
          content:
            application/json:
              example:
                success: true
                data:
                  - vehicleId: 64a1f2b3c4d5e6f7a8b9c0d5
                    registrationNumber: REG101
                    vehicleStatus: active
                    vehicleColor: Yellow
                    assignedDistrict: Colombo
                    latestAnomalyType: SPEEDING
                    anomalyTypes:
                      - SPEEDING
                      - NIGHT_MOVEMENT
                    totalAnomalyCount: 24
                    severityScore: 2
                    lastSeenAt: "2025-04-24T10:23:00.000Z"
                    lastKnownLocation:
                      latitude: 6.9271
                      longitude: 79.8612
                    lastSpeed: 95
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /alerts/suspicious/province/{provinceName}:
    get:
      summary: Get all suspicious vehicles in a province
      description: |
        Returns vehicles that have triggered anomaly flags from the
        pattern detection engine.
      operationId: getSuspiciousVehiclesByProvince
      tags:
        - Alerts
      parameters:
        - in: path
          name: provinceName
          required: true
          schema:
            type: string
          example: Western
      responses:
        '200':
          description: Suspicious vehicles in province
          content:
            application/json:
              example:
                success: true
                data:
                  - vehicleId: 64a1f2b3c4d5e6f7a8b9c0d5
                    registrationNumber: REG101
                    vehicleStatus: active
                    vehicleColor: Yellow
                    assignedProvince: Western
                    latestAnomalyType: SPEEDING
                    anomalyTypes:
                      - SPEEDING
                      - NIGHT_MOVEMENT
                    totalAnomalyCount: 24
                    severityScore: 2
                    lastSeenAt: "2025-04-24T10:23:00.000Z"
                    lastKnownLocation:
                      latitude: 6.9271
                      longitude: 79.8612
                    lastSpeed: 95
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  #master data management
...